backend/requirements.txt
Flask==2.1.1
Flask_Cors==3.0.10
marshmallow==3.15.0
mysql-connector-python==8.0.28
PyYAML==6.0
requests==2.25.1
Werkzeug==2.1.1
